+++ src/native/native.c
@@ -1,6 +1,6 @@
 #include <stdint.h>
 
-int64_t sum(int32_t a, int32_t b)
+int64_t sum(int64_t a, int64_t b)
 {
     return (a + b);
 }
